Allow opt-out of GeoJSON validation
If the client and server implementation are controlled by the same developers,
there's no need to validate the GeoJSON format since the data originates from a controlled environment.
This switch is important because we found the GeoJSON validation to be quite a performance hog
on moderate amounts of data (150 points).